refactored all files into seperate component packages
