+++ frontend/AppNavigation.js
@@ -9,36 +9,36 @@ import { createDrawerNavigator } from "@react-navigation/drawer";
 const Stack = createNativeStackNavigator();
 const Drawer = createDrawerNavigator();
 
-import { Home } from "./components/Home";
-import { Login } from "./components/Login";
-import { PasswordRecovery } from "./components/PasswordRecovery";
-import { SignUp } from "./components/SignUp";
-import { Notes } from "./components/notes-component/Notes-Main";
-import TodoScreen from "./components/TodoScreen";
-import MindMapScreen from "./components/MindMapScreen";
+import { Home } from "./components/home-component/Home";
+import { Login } from "./components/login-component/LoginScreen";
+import { PasswordRecovery } from "./components/login-component/PasswordRecoveryScreen";
+import { SignUp } from "./components/login-component/SignUpScreen";
+import { Notes } from "./components/notes-component/NotesScreen";
+import TodoScreen from "./components/todo-component/TodoScreen";
+import MindMapScreen from "./components/mindmap-component/MindMapScreen";
 
 const AppNavigation = () => {
   const auth = useContext(AuthContext);
 
   return (
     <NavigationContainer>
-      {auth.isLoggedIn ? (
+      {/* {auth.isLoggedIn ? ( */}
         <Drawer.Navigator
           screenOptions={{ headerShown: false }}
           drawerContent={CustomDrawerContent}
         >
-          <Drawer.Screen name="Home" component={Home} />
           <Drawer.Screen name="Mindmap" component={MindMapScreen} />
-          <Drawer.Screen name="Todo" component={TodoScreen} />
           <Drawer.Screen name="Notes" component={Notes} />
+          <Drawer.Screen name="Home" component={Home} />
+          <Drawer.Screen name="Todo" component={TodoScreen} />
         </Drawer.Navigator>
-      ) : (
+      {/* ) : (
         <Stack.Navigator screenOptions={{ headerShown: false }}>
           <Stack.Screen name="Login" component={Login} />
           <Stack.Screen name="Sign Up" component={SignUp} />
           <Stack.Screen name="Password Recovery" component={PasswordRecovery} />
         </Stack.Navigator>
-      )}
+      )} */}
     </NavigationContainer>
   );
 };
