+++ frontend/components/notes-component/Folder.js
@@ -2,7 +2,7 @@ import React, { useContext, useState } from "react";
 import { MaterialIcons } from '@expo/vector-icons'; 
 import { Entypo } from '@expo/vector-icons'; 
 import { AntDesign } from '@expo/vector-icons';
-import { respLgFont,respLgContHeight,respMdFont } from "../Subcomponents/CurrentFolderContainer";
+import { respLgFont,respLgContHeight,respMdFont } from "./CurrentFolderContainer";
