+++ frontend/components/notes-component/Note.js
@@ -2,7 +2,7 @@ import React, { useContext, useState } from "react";
 import { MaterialIcons } from '@expo/vector-icons'; 
 import { Entypo } from '@expo/vector-icons'; 
 import { AntDesign } from '@expo/vector-icons';
-import { respLgFont,respLgContHeight,respMdFont,respSmFont,respMiniFont } from "../Subcomponents/CurrentFolderContainer";
+import { respLgFont,respLgContHeight,respMdFont,respSmFont,respMiniFont } from "./CurrentFolderContainer";
